+++ app/__init__.py
@@ -1,3 +1,3 @@
 __version__ = "0.0.0.6"
-__asset_rev__ = "20260915l"
+__asset_rev__ = "20260915m"
 __author__ = "Adam Stone"